% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbruvs-package.R
\docType{package}
\name{rbruvs-package}
\alias{rbruvs}
\alias{rbruvs-package}
\title{rbruvs: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/frbcesab/rbruvs}
  \item Report bugs at \url{https://github.com/frbcesab/rbruvs/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Tom B. Letessier \email{tomletessier@gmail.com} (\href{https://orcid.org/0000-0003-4011-0207}{ORCID})
  \item Laura Mannocci \email{laura.mannocci@gmail.com} (\href{https://orcid.org/0000-0001-8147-8644}{ORCID})
}

}
\keyword{internal}
